+++ src/Table/Dto/Pagination/Settings.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace Patrikjak\Utils\Table\Dtos\Pagination;
+namespace Patrikjak\Utils\Table\Dto\Pagination;
 
 use Illuminate\Support\Collection;
