+++ src/Table/Services/BaseTableProvider.php
@@ -4,8 +4,8 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Services;
 
-use Patrikjak\Utils\Table\Dtos\Parameters;
-use Patrikjak\Utils\Table\Dtos\Table;
+use Patrikjak\Utils\Table\Dto\Parameters;
+use Patrikjak\Utils\Table\Dto\Table;
 
 abstract class BaseTableProvider implements TableProviderInterface
 {
