+++ src/Table/Services/TableProviderInterface.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Services;
 
-use Patrikjak\Utils\Table\Dtos\Action;
+use Patrikjak\Utils\Table\Dto\Action;
 use Patrikjak\Utils\Table\Services\ColumnTypes\Interfaces\ColumnType;
 
 interface TableProviderInterface
